requirements.txt
antlr4-python3-runtime==4.7.2
appdirs==1.4.3
autograd==1.2
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
funcsigs==1.0.2
future==0.17.1
idna==2.8
immutables==0.6
kiwisolver==1.1.0
matplotlib==3.1.0
mock==3.0.5
msgpack==0.6.1
networkx==2.3
numpy==1.16.4
PennyLane==0.3.1
pyparsing==2.4.0
pyquil==2.8.0
python-dateutil==2.8.0
python-rapidjson==0.7.2
pyzmq==18.0.1
quantum-grove==1.7.0
requests==2.22.0
rpcq==2.6.1
ruamel.yaml==0.15.97
scipy==1.3.0
semantic-version==2.6.0
six==1.12.0
toml==0.10.0
typing==3.6.6
urllib3==1.25.3
